refactor(overdue): decompose OverdueService into Analytics/Queries/LifecycleCron/Governance/ContactLog services
Behavior-preserving write-heavy (8 $transaction incl 1 Serializable; jeRefs=0).
Facade keeps 25-method surface + ctor + delegates; internally constructs 5
sub-services (specs+module+scheduler consumer untouched). 8 $tx moved WHOLE
(logContact Serializable + promiseService.createPromise(tx) passthrough intact;
partialPaymentReschedule's intentional cross-tx payment separation preserved);
raw-SQL late-fee/CTE verbatim. tsc 0, tests green, lint 0.
